reafactor(internal): move non-internal package our of internal
As the design iterates, only context along with test has internal implementations.
Only unexposed functions uses context.ResponseChannel to communicate with exposed functions.
So there is no need to unexpose package command, gitop, k8sop, and slashcommand.
