+++ command/info.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/mirror-media/major-tom-go/config"
-	"github.com/mirror-media/major-tom-go/internal/k8sop"
+	"github.com/mirror-media/major-tom-go/k8sop"
 	"github.com/pkg/errors"
 )
