+++ command/info_test.go
